chore: 스타일 코드 디렉토리 및 파일 분리(#120)
