+++ src/components/search/searchFound/SearchFound.utils.ts
@@ -1,0 +1,9 @@
+import { IGenerations } from '@/types/request';
+
+export const getGenerationFormat = (generations: IGenerations[]) => {
+	const genFormat: string[] = [];
+	generations.forEach((data) => {
+		genFormat.push(data.generation);
+	});
+	return genFormat.join(', ');
+};
