+++ src/components/search/searchHeaderForm/SearchHeaderForm.styled.ts
@@ -1,0 +1,26 @@
+import styled from 'styled-components';
+
+export const FormWrapper = styled.form``;
+
+export const SearchWrapper = styled.div`
+	display: flex;
+	align-items: center;
+	border-bottom: 1px solid black;
+	padding: 2px 0rem;
+`;
+
+export const SearchHeaderInput = styled.input`
+	font-family: NeoDunggeunmo Pro;
+	width: 100%;
+	height: 100%;
+	font-size: 1.5rem;
+	border: none;
+	outline: none;
+	padding-left: 1rem;
+`;
+
+export const ImageButtonWrapper = styled.button`
+	border: none;
+	outline: none;
+	background-color: inherit;
+`;
